+++ pyrr/tests/test_euler.py
@@ -5,6 +5,7 @@ except:
 import numpy as np
 from pyrr import euler
 
+
 class test_euler(unittest.TestCase):
     def test_import(self):
         import pyrr
@@ -12,20 +13,18 @@ class test_euler(unittest.TestCase):
         from pyrr import euler
 
     def test_create(self):
-        self.assertTrue(np.array_equal(euler.create(), [0.,0.,0.]))
-        self.assertTrue(np.array_equal(euler.create(1.,2.,3.), [1.,2.,3.]))
+        self.assertTrue(np.array_equal(euler.create(), [0., 0., 0.]))
+        self.assertTrue(np.array_equal(euler.create(roll=1., pitch=2., yaw=3.), [1., 2., 3.]))
 
-    def test_pitch(self):
-        self.assertEqual(euler.pitch([1.,2.,3.]), 1.)
+        # swap indices
+        indices = euler.index.yaw, euler.index.pitch, euler.index.roll
+        new_indices = indices[1], indices[2], indices[0]
+        euler.index.yaw, euler.index.pitch, euler.index.roll = new_indices
 
-    def test_roll(self):
-        self.assertEqual(euler.roll([1.,2.,3.]), 2.)
+        self.assertTrue(np.array_equal(euler.create(roll=1., pitch=2., yaw=3.), [2., 3., 1.]))
 
-    def test_yaw(self):
-        self.assertEqual(euler.yaw([1.,2.,3.]), 3.)
-
-
+        # reset
+        euler.index.yaw, euler.index.pitch, euler.index.roll = indices
 
 if __name__ == '__main__':
     unittest.main()
-